chore: Add isAdmin middleware for admin authorization
This commit adds the `isAdmin` middleware to the `authenticate.ts` file. The `isAdmin` middleware checks if the user is an admin by querying the `userModel` with the `userId` obtained from the JWT token. If the user is an admin, the next middleware is called. Otherwise, an unauthorized error is returned.

Closes #123